pyproject.toml
[project]
name = "contentunderstanding"
version = "0.1.0"
description = "Document Validation System - Local-First POC"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "python-dateutil>=2.8",
    "openai>=1.0",
    "azure-ai-documentintelligence>=1.0.0b4",
    "azure-identity>=1.15",
    "azure-storage-blob>=12.19",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
